requirements.txt
pymongo==4.4.1
python-dotenv
streamlit==1.29.0